requirements.txt
certifi==2018.8.24
chardet==3.0.4
click==6.7
Django==2.1.1
django-crispy-forms==1.7.2
djangorestframework==3.10.2
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Markdown==3.1
MarkupSafe==1.0
oauthlib==2.1.0
requirements==0.1
urllib3==1.23
virtualenv==15.2.0
Werkzeug==0.14.1
whitenoise==3.3.1
